Share the test fixtures between the indexing tests
Copies of the same helpers had grown as each fix added a file: toLogical in
two, and the array builders in one that a third test now needs. Move them to
__tests__/helpers.ts.
